Added props and types for AmountInput to Card Components
