+++ src/components/atoms/AmountInput/index.ts
@@ -1,0 +1,2 @@
+export * from './AmountInput';
+export * from './AmountInput.types';
