+++ src/components/atoms/Avatar/index.ts
@@ -1,0 +1,3 @@
+export * from './Avatar';
+export * from './Avatar.types';
+
