+++ src/components/atoms/Badge/Badge.tsx
@@ -1,28 +1,8 @@
 import React from 'react';
-import { ColorType } from '../../../types/colors';
 import { Avatar } from '../Avatar/Avatar';
-import { Icon, IconType } from '../icon';
+import { BadgeProps, relevantStyleToSize } from './Badge.types';
+import { Icon } from '../icon';
 import { Typography } from '../Typography';
-
-/**
- * @params color - Text color for badge with color/20 background
- * @params label - Text within badge
- */
-export interface BadgeProps extends React.HTMLAttributes<HTMLDivElement> {
-  avatar?: string;
-  className?: string;
-  color?: Exclude<ColorType, 'white' | 'black'>;
-  endIcon?: IconType;
-  icon?: IconType;
-  label: string;
-  size?: 'sm' | 'md' | 'lg';
-  status?: 'online' | 'offline' | 'away';
-}
-const relevantStyleToSize = {
-  sm: 'bdge-sm',
-  md: 'bdge-md',
-  lg: 'bdge-lg',
-};
 
 export function Badge({
   label,
