+++ src/components/atoms/Badge/index.ts
@@ -1,0 +1,2 @@
+export * from './Badge';
+export * from './Badge.types';
