+++ src/components/atoms/ButtonGroup/ButtonGroup.types.ts
@@ -1,6 +1,5 @@
-import { HTMLAttributes } from 'react';
-import { PrimaryColorType } from '../../../types/colors';
-import { IconType } from '../icon';
+import { ButtonProps } from "../Button/Button.types";
+
 /**
  * @params EndIcon - The icon to be displayed at the end of the button
  * @params StartIcon - The icon to be displayed at the start of the button
@@ -12,30 +11,11 @@ import { IconType } from '../icon';
  * @params shape - Shape of the button. It supports those theme shapes that make sense for this component.
  * @params size - Size of the button. It supports those theme sizes that make sense for this component.
  * @params variant - Variant of the button. It supports those theme variants that make sense for this component.
+ * @params orientation - Horizontal/Vertical Orientation of button group.
  */
-export interface ButtonProps extends HTMLAttributes<HTMLButtonElement | HTMLAnchorElement> {
+export interface ButtonGroupProps
+  extends Pick<ButtonProps, 'color' | 'size' | 'variant' | 'icon' | 'endIcon'> {
   className?: string;
-  color?: Exclude<PrimaryColorType, 'white' | 'black'>;
-  disabled?: boolean;
-  endIcon?: IconType;
-  href?: string;
-  icon?: IconType;
-  label: string;
-  online?: boolean;
-  selected?: boolean;
-  shape?: 'rounded' | 'square';
-  size?: 'sm' | 'md' | 'lg' | 'xl' | '2xl';
-  textAlign?: 'left' | 'center' | 'right';
-  variant?: 'contained' | 'outlined' | 'light' | 'text' | 'link';
+  items: Array<Omit<ButtonProps, 'color' | 'size' | 'variant' | 'shape'>>;
+  orientation?: 'horizontal' | 'vertical';
 }
-export const sizes = {
-  sm: 'btn-sm',
-  md: 'btn-md',
-  lg: 'btn-lg',
-  xl: 'btn-xl',
-  '2xl': 'btn-2xl',
-};
-export const shapes = {
-  rounded: 'rounded-full',
-  square: 'rounded',
-};
