+++ src/components/atoms/ButtonGroup/index.ts
@@ -1,0 +1,2 @@
+export * from './ButtonGroup'
+export * from './ButtonGroup.types'
