+++ src/components/atoms/ButtonWrapper/index.ts
@@ -1,0 +1,2 @@
+export * from './ButtonWrapper'
+export * from './ButtonWrapper.types'
