+++ src/components/atoms/card/index.ts
@@ -1,0 +1,2 @@
+export * from './card'
+export * from './card.types'
