+++ src/components/atoms/checkbox/index.ts
@@ -1,0 +1,2 @@
+export * from './checkbox'
+export * from './checkbox.types'
